4.21*: blogilistan laajennus, step9
Poisto sallittu vain blogin omistajalle

+++ osa4/blogilista/controllers/blogs.js
@@ -1,19 +1,19 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const jwt = require("jsonwebtoken");
+const middleware = require("../utils/middleware");
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {username: 1, name: 1, id: 1})
   response.json(blogs)
 })
 
+// APIS REQUIRING TOKEN
+blogsRouter.use(middleware.tokenExtractor)
+
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
-  if (request.token === null || request.token.id === undefined) {
-    return response.status(401).json({ error: 'token required' })
-  }
   const user = await User.findById(request.token.id)
   if (user === null) {
     return response.status(401).json({ error: 'user not found'})
@@ -33,6 +33,16 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const user = await User.findById(request.token.id)
+  if (user === null) {
+    return response.status(401).json({ error: 'user not found'})
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(403).json({ error: 'Won\'t delete - not your blog!'})
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
